dashboard/backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
notion-client==2.2.1
pydantic==2.5.3
python-dotenv==1.0.0
